# code/backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateUserRequest:
    properties:
      password:
        type: string
      profile_picture:
        type: string
      username:
        type: string
    type: object
...
  models.GameSummary:
    properties:
      gameId:
        type: string
      gameType:
        type: string
      isGameOver:
        type: boolean
      isRunning:
        type: boolean
      isSetupPhase:
        type: boolean
      player1Username:
        type: string
      player2Username:
        type: string
      round:
        type: integer
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      profile_picture:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: info@dotsem.be
    name: Sem Van Broekhoven
    url: https://github.com/Thomas-More-Digital-Innovation/2526-DI-004-GoStrategy
  description: This is the API server for the GoStrategy game.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  termsOfService: http://swagger.io/terms/
  title: GoStrategy API
  version: 0.1.1
paths:
...
  /game/{gameID}:
    get:
      description: Establish a WebSocket connection to an active game session
      parameters:
      - description: Game ID
        in: path
        name: gameID
        required: true
        type: string
      - description: Player index (0 or 1)
        in: query
        name: player
        type: string
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: Game ID required or invalid player index
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Game not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Connect to game via WebSocket
      tags:
      - games
  /games:
    get:
      description: Retrieve a list of all currently active game sessions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GameSummary'
            type: array
      summary: List active games
      tags:
      - games
    post:
      consumes:
      - application/json
      description: Initialize a new game session with specified type and players
      parameters:
      - description: Game creation parameters
        in: body
        name: request
        required: true
        schema:
          properties:
            ai1:
              type: string
            ai2:
              type: string
            gameId:
              type: string
            gameType:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Game created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or parameters
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new game
      tags:
      - games
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return user details with session cookies
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid username or password
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
...
  /users/refresh:
    post:
      description: Renew the access token using a valid refresh token from cookies
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Refresh token missing or invalid
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with username and password
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request body or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Username already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create user
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
swagger: "2.0"